refactor:fix the no-unescaped entities in JSX error
