+++ logify-frontend/src/components/NotFound.jsx
@@ -16,7 +16,7 @@ const NotFound = () => {
           Page Not Found
         </h2>
         <p className="text-xl text-gray-600 mb-8 max-w-md">
-          Sorry, the page you're looking for doesn't exist. It might have been
+          Sorry, the page you&apos;re looking for doesn&apos;t exist. It might have been
           moved or deleted.
         </p>
