+++ logify-frontend/src/pages/dashboards/InternshipAdmin/App1.jsx
@@ -7,7 +7,7 @@ import Placements from "./pages/Placements.jsx";
 import Evaluations from "./pages/Evaluations.jsx";
 import Reports from "./pages/Reports.jsx";
 import Settings from "./pages/Settings.jsx";
-import NotFound from "./pages/NotFound.jsx";
+import NotFound from "../../../components/NotFound.jsx";
 import Sidebar from "./Sidebar.jsx";
 
 const App1 = () => {
@@ -25,7 +25,7 @@ const App1 = () => {
             <Route path="/evaluations" element={<Evaluations />} />
             <Route path="/reports" element={<Reports />} />
             <Route path="/settings" element={<Settings />} />
-            <Route path="*" element={<NotFound />} />
+            <Route path="/notfound" element={<NotFound />} />
           </Routes>
         </main>
       </div>
